Add basic internal notification system (#110)
* Add ability to create notifications

* Show notifications count in drawer

- Also add notification when creating a post

* Make notifications part of...notifications

I promise the main message makes sense. Before this, we were dealing
with notifications ouside of that genserver. I'm not sure that this is
the right abstraction, because in-app notifications have different
batching concerns than emails do, but this makes more sense than what I
was doing before this commit!